renamed community data server from data.server
