+++ app/routes/admin/programs/mockData.ts
@@ -1,12 +1,14 @@
 import type { Enrollee } from "./program_tables";
 
+type StatusType = "submitted" | "accepted" | "waitlist" | "declined";
+
 export const mockEnrollees: Enrollee[] = [
   {
     id: "1",
     firstName: "Leslie",
     lastName: "Alexander",
     email: "leslie.alexander@example.com",
-    status: "enrolled",
+    status: "submitted",
     enrollmentDate: "Aug. 15 2025",
     appliedDate: "Aug. 1 2025",
     image: "https://randomuser.me/api/portraits/women/1.jpg",
@@ -16,7 +18,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Michael",
     lastName: "Smith",
     email: "michael.smith@example.com",
-    status: "pending",
+    status: "accepted",
     enrollmentDate: "Aug. 16 2025",
     appliedDate: "Aug. 2 2025",
     image: "https://randomuser.me/api/portraits/men/2.jpg",
@@ -26,7 +28,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Emily",
     lastName: "Johnson",
     email: "emily.johnson@example.com",
-    status: "enrolled",
+    status: "waitlist",
     enrollmentDate: "Aug. 17 2025",
     appliedDate: "Aug. 3 2025",
     image: "https://randomuser.me/api/portraits/women/3.jpg",
@@ -36,7 +38,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "David",
     lastName: "Williams",
     email: "david.williams@example.com",
-    status: "failed",
+    status: "declined",
     enrollmentDate: "Aug. 18 2025",
     appliedDate: "Aug. 4 2025",
     image: "https://randomuser.me/api/portraits/men/4.jpg",
@@ -46,7 +48,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Sarah",
     lastName: "Brown",
     email: "sarah.brown@example.com",
-    status: "pending",
+    status: "submitted",
     enrollmentDate: "Aug. 19 2025",
     appliedDate: "Aug. 5 2025",
     image: "https://randomuser.me/api/portraits/women/5.jpg",
@@ -56,7 +58,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "James",
     lastName: "Jones",
     email: "james.jones@example.com",
-    status: "enrolled",
+    status: "accepted",
     enrollmentDate: "Aug. 20 2025",
     appliedDate: "Aug. 6 2025",
     image: "https://randomuser.me/api/portraits/men/6.jpg",
@@ -66,7 +68,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Olivia",
     lastName: "Garcia",
     email: "olivia.garcia@example.com",
-    status: "pending",
+    status: "waitlist",
     enrollmentDate: "Aug. 21 2025",
     appliedDate: "Aug. 7 2025",
     image: "https://randomuser.me/api/portraits/women/7.jpg",
@@ -76,7 +78,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "William",
     lastName: "Martinez",
     email: "william.martinez@example.com",
-    status: "enrolled",
+    status: "declined",
     enrollmentDate: "Aug. 22 2025",
     appliedDate: "Aug. 8 2025",
     image: "https://randomuser.me/api/portraits/men/8.jpg",
@@ -86,7 +88,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Sophia",
     lastName: "Hernandez",
     email: "sophia.hernandez@example.com",
-    status: "failed",
+    status: "submitted",
     enrollmentDate: "Aug. 23 2025",
     appliedDate: "Aug. 9 2025",
     image: "https://randomuser.me/api/portraits/women/9.jpg",
@@ -96,7 +98,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Benjamin",
     lastName: "Lopez",
     email: "benjamin.lopez@example.com",
-    status: "pending",
+    status: "accepted",
     enrollmentDate: "Aug. 24 2025",
     appliedDate: "Aug. 10 2025",
     image: "https://randomuser.me/api/portraits/men/10.jpg",
@@ -106,7 +108,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Ava",
     lastName: "Gonzalez",
     email: "ava.gonzalez@example.com",
-    status: "enrolled",
+    status: "waitlist",
     enrollmentDate: "Aug. 25 2025",
     appliedDate: "Aug. 11 2025",
     image: "https://randomuser.me/api/portraits/women/11.jpg",
@@ -116,7 +118,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Lucas",
     lastName: "Wilson",
     email: "lucas.wilson@example.com",
-    status: "pending",
+    status: "declined",
     enrollmentDate: "Aug. 26 2025",
     appliedDate: "Aug. 12 2025",
     image: "https://randomuser.me/api/portraits/men/12.jpg",
@@ -126,7 +128,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Mia",
     lastName: "Anderson",
     email: "mia.anderson@example.com",
-    status: "enrolled",
+    status: "submitted",
     enrollmentDate: "Aug. 27 2025",
     appliedDate: "Aug. 13 2025",
     image: "https://randomuser.me/api/portraits/women/13.jpg",
@@ -136,7 +138,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Henry",
     lastName: "Thomas",
     email: "henry.thomas@example.com",
-    status: "failed",
+    status: "accepted",
     enrollmentDate: "Aug. 28 2025",
     appliedDate: "Aug. 14 2025",
     image: "https://randomuser.me/api/portraits/men/14.jpg",
@@ -146,7 +148,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Charlotte",
     lastName: "Taylor",
     email: "charlotte.taylor@example.com",
-    status: "pending",
+    status: "waitlist",
     enrollmentDate: "Aug. 29 2025",
     appliedDate: "Aug. 15 2025",
     image: "https://randomuser.me/api/portraits/women/15.jpg",
@@ -156,7 +158,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Jack",
     lastName: "Moore",
     email: "jack.moore@example.com",
-    status: "enrolled",
+    status: "declined",
     enrollmentDate: "Aug. 30 2025",
     appliedDate: "Aug. 16 2025",
     image: "https://randomuser.me/api/portraits/men/16.jpg",
@@ -166,7 +168,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Amelia",
     lastName: "Jackson",
     email: "amelia.jackson@example.com",
-    status: "pending",
+    status: "submitted",
     enrollmentDate: "Aug. 31 2025",
     appliedDate: "Aug. 17 2025",
     image: "https://randomuser.me/api/portraits/women/17.jpg",
@@ -176,7 +178,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Ethan",
     lastName: "Martin",
     email: "ethan.martin@example.com",
-    status: "enrolled",
+    status: "accepted",
     enrollmentDate: "Sep. 1 2025",
     appliedDate: "Aug. 18 2025",
     image: "https://randomuser.me/api/portraits/men/18.jpg",
@@ -186,7 +188,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Harper",
     lastName: "Lee",
     email: "harper.lee@example.com",
-    status: "failed",
+    status: "waitlist",
     enrollmentDate: "Sep. 2 2025",
     appliedDate: "Aug. 19 2025",
     image: "https://randomuser.me/api/portraits/women/19.jpg",
@@ -196,7 +198,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Alexander",
     lastName: "Perez",
     email: "alexander.perez@example.com",
-    status: "pending",
+    status: "declined",
     enrollmentDate: "Sep. 3 2025",
     appliedDate: "Aug. 20 2025",
     image: "https://randomuser.me/api/portraits/men/20.jpg",
@@ -206,7 +208,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Ella",
     lastName: "White",
     email: "ella.white@example.com",
-    status: "enrolled",
+    status: "submitted",
     enrollmentDate: "Sep. 4 2025",
     appliedDate: "Aug. 21 2025",
     image: "https://randomuser.me/api/portraits/women/21.jpg",
@@ -216,7 +218,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Daniel",
     lastName: "Harris",
     email: "daniel.harris@example.com",
-    status: "pending",
+    status: "accepted",
     enrollmentDate: "Sep. 5 2025",
     appliedDate: "Aug. 22 2025",
     image: "https://randomuser.me/api/portraits/men/22.jpg",
@@ -226,7 +228,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Grace",
     lastName: "Clark",
     email: "grace.clark@example.com",
-    status: "enrolled",
+    status: "waitlist",
     enrollmentDate: "Sep. 6 2025",
     appliedDate: "Aug. 23 2025",
     image: "https://randomuser.me/api/portraits/women/23.jpg",
@@ -236,7 +238,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Matthew",
     lastName: "Lewis",
     email: "matthew.lewis@example.com",
-    status: "failed",
+    status: "declined",
     enrollmentDate: "Sep. 7 2025",
     appliedDate: "Aug. 24 2025",
     image: "https://randomuser.me/api/portraits/men/24.jpg",
@@ -246,7 +248,7 @@ export const mockEnrollees: Enrollee[] = [
     firstName: "Chloe",
     lastName: "Walker",
     email: "chloe.walker@example.com",
-    status: "pending",
+    status: "submitted",
     enrollmentDate: "Sep. 8 2025",
     appliedDate: "Aug. 25 2025",
     image: "https://randomuser.me/api/portraits/women/25.jpg",
